User required for image repo access

+++ www/api/repositories/ImageRepository.php
@@ -1,5 +1,6 @@
 <?php
 
+require_once __DIR__ . '/../models/Album.php';
 require_once __DIR__ . '/../models/Image.php';
 
 class ImageRepository {
@@ -28,7 +29,7 @@ class ImageRepository {
      * Returns an Image object or false if creation fails.
      * This method can only be called from within the class on successful upload.
      */
-    public function createImage($id, $extension, $imageCaption, $imageType, $imageSize, $imageUploadDate, $userId = null) {
+    public function createImage($id, $extension, $imageCaption, $imageType, $imageSize, $imageUploadDate, $userId) {
         // Sanitize user input to prevent SQL injection
         $id = htmlspecialchars($id);
         $extension = htmlspecialchars($extension);
@@ -50,30 +51,89 @@ class ImageRepository {
         $stmt->bindParam(':userId', $userId);
 
         if ($stmt->execute()) {
-            // User registration successful
+            // Image creation successful
             return new Image($id, $extension, $imageCaption, $imageType, $imageSize, $imageUploadDate, $userId);
         } else {
-            // User registration failed
+            // Image creation failed
             return false;
         }
     }
 
     /*
-     * Returns an Image object or false if the image is not found
+     * Returns an Album object or false if creation fails.
      */
-    public function getImageById($id) {
+    public function createAlbum($albumName, $description, $userId) {
         // Sanitize user input to prevent SQL injection
-        $id = htmlspecialchars($id);
+        $albumName = htmlspecialchars($albumName);
+        $description = htmlspecialchars($description);
+        $userId = htmlspecialchars($userId);
+
+        // Insert new album into the database
+        $insertImageQuery = "INSERT INTO albums (album_name, description, user_id) VALUES (:albumName, :description, :userId)";
+        $stmt = $this->db->prepare($insertImageQuery);
+        $stmt->bindParam(':albumName', $albumName);
+        $stmt->bindParam(':description', $description);
+        $stmt->bindParam(':userId', $userId);
+
+        if ($stmt->execute()) {
+            // Album creation successful
+            return new Image($albumName, $description, $userId);
+        } else {
+            // Album creation failed
+            return false;
+        }
+    }
+
+    /*
+     * Returns an album or false if the album is not found
+     */
+    public function getAlbumByName($albumName, $userId) {
+        // Sanitize user input to prevent SQL injection
+        $albumName = htmlspecialchars($albumName);
+        $userId = htmlspecialchars($userId);
 
         // Get image from the database
-        $getImageQuery = "SELECT * FROM images WHERE id = :id";
+        $getAlbumQuery = "SELECT * FROM images WHERE album_name = :albumName AND user_id = :userId";
+        $stmt = $this->db->prepare($getAlbumQuery);
+        $stmt->bindParam(':albumName', $albumName);
+        $stmt->bindParam(':userId', $userId);
+        $stmt->execute();
+        $album = $stmt->fetch(PDO::FETCH_ASSOC);
+
+        if ($album) {
+            // album found
+            return new Album($album['album_name'], $album['description'], $album['user_id']);
+        } else {
+            // album not found
+            return false;
+        }
+    }
+
+    /*
+     * Returns an Image object/Image array or false if the image(s) is not found
+     */
+    public function getImageById($ids, $userId) {
+        // Sanitize user input to prevent SQL injection
+        $id = htmlspecialchars($ids);
+
+        // Get image from the database
+        $getImageQuery = "SELECT * FROM images WHERE id IN(:id) AND user_id = :userId";
         $stmt = $this->db->prepare($getImageQuery);
         $stmt->bindParam(':id', $id);
+        $stmt->bindParam(':userId', $userId);
         $stmt->execute();
-        $image = $stmt->fetch(PDO::FETCH_ASSOC);
+        $images = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
-        if ($image) {
-            // Image found
+        if ($images) {
+            // Image(s) found
+            if (count($images) > 1) {
+                $imageObjects = array();
+                foreach ($images as $image) {
+                    $imageObjects[] = new Image($image['id'], $image['extension'], $image['image_caption'], $image['image_type'], $image['image_size'], $image['image_upload_date'], $image['user_id']);
+                }
+                return $imageObjects;
+            }
+            $image = $images[0];
             return new Image($image['id'], $image['extension'], $image['image_caption'], $image['image_type'], $image['image_size'], $image['image_upload_date'], $image['user_id']);
         } else {
             // Image not found
@@ -111,15 +171,12 @@ class ImageRepository {
     /*
      * Returns an array of Image objects or false if no images are found
      */
-    // TODO: add params for pagination of large albums
-    public function getImagesForAlbum($albumName) {
-        // Sanitize user input to prevent SQL injection
-        $albumName = htmlspecialchars($albumName);
-
+    // TODO: add params for pagination
+    public function getAllImages($userId) {
         // Get images from the database
-        $getImagesQuery = "SELECT * FROM images AS i JOIN album_images AS a ON i.id = a.image_id WHERE a.albumName = :albumName";
+        $getImagesQuery = "SELECT * FROM images WHERE user_id = :userId";
         $stmt = $this->db->prepare($getImagesQuery);
-        $stmt->bindParam(':albumName', $albumName);
+        $stmt->bindParam(':userId', $userId);
         $stmt->execute();
         $images = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
@@ -127,7 +184,7 @@ class ImageRepository {
             // Images found
             $imageObjects = array();
             foreach ($images as $image) {
-                $imageIds[] = new Image($image['id'], $image['extension'], $image['image_caption'], $image['image_type'], $image['image_size'], $image['image_upload_date'], $image['user_id']);
+                $imageObjects[] = new Image($image['id'], $image['extension'], $image['image_caption'], $image['image_type'], $image['image_size'], $image['image_upload_date'], $image['user_id']);
             }
             return $imageObjects;
         } else {
@@ -137,30 +194,85 @@ class ImageRepository {
     }
 
     /*
-     * Returns an Image object or false if the image is not found
+     * Returns an array of Album objects or false if no albums are found
      */
-    public function updateImage($image) {
+    // TODO: add params for pagination
+    public function getAllAlbums($userId) {
+        // Get albums from the database
+        $getAlbumsQuery = "SELECT * FROM albums WHERE user_id = :userId";
+        $stmt = $this->db->prepare($getAlbumsQuery);
+        $stmt->bindParam(':userId', $userId);
+        $stmt->execute();
+        $albums = $stmt->fetchAll(PDO::FETCH_ASSOC);
+
+        if ($albums) {
+            // albums found
+            $albumObjects = array();
+            foreach ($albums as $image) {
+                $albumObjects[] = new Album($image['album_name'], $image['description'], $image['user_id']);
+            }
+            return $albumObjects;
+        } else {
+            // No images not found
+            return false;
+        }
+    }
+
+    /*
+     * Returns an array of Image objects or false if no images are found
+     */
+    // TODO: add params for pagination
+    public function getImagesForAlbum($albumName, $userId) {
         // Sanitize user input to prevent SQL injection
-        $filename = htmlspecialchars($image->filename);
+        $albumName = htmlspecialchars($albumName);
+        $userId = htmlspecialchars($userId);
+
+        // Get images from the database
+        $getImagesQuery = "SELECT * FROM images AS i JOIN album_images AS a ON i.id = a.image_id WHERE a.album_name = :albumName AND i.user_id = :userId";
+        $stmt = $this->db->prepare($getImagesQuery);
+        $stmt->bindParam(':albumName', $albumName);
+        $stmt->bindParam(':userId', $userId);
+        $stmt->execute();
+        $images = $stmt->fetchAll(PDO::FETCH_ASSOC);
+
+        if ($images) {
+            // Images found
+            $imageObjects = array();
+            foreach ($images as $image) {
+                $imageObjects[] = new Image($image['id'], $image['extension'], $image['image_caption'], $image['image_type'], $image['image_size'], $image['image_upload_date'], $image['user_id']);
+            }
+            return $imageObjects;
+        } else {
+            // No images not found
+            return false;
+        }
+    }
+
+    /*
+     * Returns an true if the image was updated, false otherwise
+     */
+    public function updateImage($image, $userId) {
+        // Sanitize user input to prevent SQL injection
+        $id = htmlspecialchars($image->id);
         $imageCaption = htmlspecialchars($image->imageCaption);
-        $imageType = htmlspecialchars($image->imageType);
-        $imageSize = htmlspecialchars($image->imageSize);
-        $imageUploadDate = htmlspecialchars($image->imageUploadDate);
+        // $imageType = htmlspecialchars($image->imageType);
+        // $imageSize = htmlspecialchars($image->imageSize);
+        // $imageUploadDate = htmlspecialchars($image->imageUploadDate);
         $userId = htmlspecialchars($image->userId);
 
         // Update image in the database
-        $updateImageQuery = "UPDATE images SET image_caption = :imageCaption, image_type = :imageType, image_size = :imageSize, image_upload_date = :imageUploadDate, user_id = :userId WHERE filename = :filename";
+        $updateImageQuery = "UPDATE images SET image_caption = :imageCaption WHERE id = :id AND user_id = :userId"; //, image_type = :imageType, image_size = :imageSize, image_upload_date = :imageUploadDate, user_id = :userId 
         $stmt = $this->db->prepare($updateImageQuery);
-        $stmt->bindParam(':filename', $filename);
+        $stmt->bindParam(':id', $id);
         $stmt->bindParam(':imageCaption', $imageCaption);
-        $stmt->bindParam(':imageType', $imageType);
-        $stmt->bindParam(':imageSize', $imageSize);
-        $stmt->bindParam(':imageUploadDate', $imageUploadDate);
+        // $stmt->bindParam(':imageType', $imageType);
+        // $stmt->bindParam(':imageSize', $imageSize);
+        // $stmt->bindParam(':imageUploadDate', $imageUploadDate);
         $stmt->bindParam(':userId', $userId);
 
         if ($stmt->execute()) {
             // Image update successful
-            return new Image($filename, $imageCaption, $imageType, $imageSize, $imageUploadDate, $userId);
+            return true; //new Image($filename, $imageCaption, $imageType, $imageSize, $imageUploadDate, $userId);
         } else {
             // Image update failed
             return false;
@@ -171,11 +283,19 @@ class ImageRepository {
      * Returns true if the image is deleted or false if the image is not found.
      * Deletes the file from the server and the database.
      */
-    public function deleteImage($id) {
+    public function deleteImage($ids, $userId) {
         // Delete image from the database
-        $deleteImageQuery = "DELETE FROM images WHERE id = :id";
+        if (is_array($ids)) {
+            $ids = array_map('htmlspecialchars', $ids);
+            $ids = implode(',', $ids);
+        } else {
+            $ids = htmlspecialchars($ids);
+        }
+
+        $deleteImageQuery = "DELETE FROM images WHERE id IN(:ids) AND user_id = :userId";
         $stmt = $this->db->prepare($deleteImageQuery);
-        $stmt->bindParam(':id', $id);
+        $stmt->bindParam(':ids', $ids);
+        $stmt->bindParam(':userId', $userId);
         return $stmt->execute();
     }
 
@@ -183,25 +303,27 @@ class ImageRepository {
      * Returns true if the images are deleted or false if no images are found.
      * Deletes the files from the server and the database.
      */
-    public function deleteImagesFromAlbum($albumName, $fromStorage = false) {
+    public function deleteImagesFromAlbum($albumName, $userId, $fromStorage = false) {
         // Sanitize user input to prevent SQL injection
         $albumName = htmlspecialchars($albumName);
         $fromStorage = !!$fromStorage;
 
         if (!$fromStorage) {
             // Delete images from the database album_images table
-            $deleteImagesQuery = "DELETE FROM album_images WHERE album_name = :albumName";
+            $deleteImagesQuery = "DELETE FROM album_images as ai INNER JOIN albums as a using(album_name) WHERE ai.album_name = :albumName AND a.user_id = :userId";
             $stmt = $this->db->prepare($deleteImagesQuery);
             $stmt->bindParam(':albumName', $albumName);
+            $stmt->bindParam(':userId', $userId);
             $stmt->execute();
 
             // Images deleted from album successful
             return true;
         } else {
             // Get image filenames from the database
-            $getImageFilenamesQuery = "SELECT id FROM album_images WHERE album_name = :albumName";
+            $getImageFilenamesQuery = "SELECT ai.image_id FROM album_images as ai INNER JOIN albums as a using(album_name) WHERE ai.album_name = :albumName AND a.user_id = :userId";
             $stmt = $this->db->prepare($getImageFilenamesQuery);
             $stmt->bindParam(':albumName', $albumName);
+            $stmt->bindParam(':userId', $userId);
             $stmt->execute();
             $imageFilenames = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
@@ -211,28 +333,30 @@ class ImageRepository {
                     $imagePath = $this->uploadPath . $imageFilename['filename'];
 
                     if (file_exists($imagePath)) {
-                        if (unlink($imagePath)) {
-                            // Delete images from any and all albums in the database
-                            $deleteImagesQuery = "DELETE FROM album_images WHERE filename = :filename";
-                            $stmt = $this->db->prepare($deleteImagesQuery);
-                            $stmt->bindParam(':filename', $imageFilename['filename']);
-                            $stmt->execute();
-                        } else {
+                        if (!unlink($imagePath)) {
                             return false;
-                        }
-                    } else {
-                        // File doesn't exist
-                        // Delete images from any and all albums in the database
-                        $deleteImagesQuery = "DELETE FROM album_images WHERE filename = :filename";
-                        $stmt = $this->db->prepare($deleteImagesQuery);
-                        $stmt->bindParam(':filename', $imageFilename['filename']);
-                        $stmt->execute();
-                    }
+                            // Delete images from any and all albums in the database
+                            // $deleteImagesQuery = "DELETE FROM album_images WHERE filename = :filename";
+                            // $stmt = $this->db->prepare($deleteImagesQuery);
+                            // $stmt->bindParam(':filename', $imageFilename['filename']);
+                            // $stmt->execute();
+                        } // else {
+                        //     return false;
+                        // }
+                    } // else {
+                    //     // File doesn't exist
+                    //     // Delete images from any and all albums in the database
+                    //     $deleteImagesQuery = "DELETE FROM album_images WHERE filename = :filename";
+                    //     $stmt = $this->db->prepare($deleteImagesQuery);
+                    //     $stmt->bindParam(':filename', $imageFilename['filename']);
+                    //     $stmt->execute();
+                    // }
 
                     // Delete image from the database
-                    $deleteImagesQuery = "DELETE FROM images WHERE filename = :filename";
+                    $deleteImagesQuery = "DELETE FROM images WHERE filename = :filename AND user_id = :userId";
                     $stmt = $this->db->prepare($deleteImagesQuery);
                     $stmt->bindParam(':filename', $imageFilename['filename']);
+                    $stmt->bindParam(':userId', $userId);
                     $stmt->execute();
 
                     // Deleted all images
